requirements.txt
pandas
numpy
matplotlib
scipy
streamlit
scikit-learn
pymysql
plotly
python-dotenv
statsmodels
fbprophet
tensorflow